Convert NanoMethResult object to edgeR methylation matrix

Convert NanoMethResult object to edgeR methylation matrix

Usage

methy_to_edger(methy, regions = NULL, out_folder = tempdir(), verbose = TRUE)

Arguments

methy

the NanoMethResult object or path to the methylation tabix file.

regions

the regions to calculate log-methylation ratios over. If left NULL, ratios will be calculated per site.

out_folder

the folder to store intermediate files. One file is created for each sample and contains columns "chr", "pos", "total" and "methylated".

verbose

TRUE if progress messages are to be printed

Value

a matrix compatible with the edgeR differential methylation pipeline

Examples

nmr <- load_example_nanomethresult()
edger_mat <- methy_to_edger(nmr)
#> [2025-05-04 11:02:40] creating intermediate files...
#> [2025-05-04 11:02:40] parsing chr11...
#> [2025-05-04 11:02:40] parsing chr12...
#> [2025-05-04 11:02:40] parsing chr18...
#> [2025-05-04 11:02:40] parsing chr5...
#> [2025-05-04 11:02:40] parsing chr7...
#> [2025-05-04 11:02:40] parsing chrX...
#> [2025-05-04 11:02:40] samples found: B6Cast_Prom_3_cast B6Cast_Prom_3_bl6 B6Cast_Prom_2_cast B6Cast_Prom_2_bl6 B6Cast_Prom_1_cast B6Cast_Prom_1_bl6 
#> [2025-05-04 11:02:40] creating bsseq object...
#> [2025-05-04 11:02:40] reading in parsed data...
#> [2025-05-04 11:02:40] constructing matrices...
#> [2025-05-04 11:02:40] done